requirements.txt
asgiref==3.6.0
Django==3.2.18
django-appconf==1.0.5
django-crispy-forms==2.0
django-imagekit==4.1.0
django-taggit==4.0.0
pilkit==2.0
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
